feat(phase B): wire play flow to the local store (instant gameplay)
Gameplay actions now run the pure engine client-side against the local store
instead of round-tripping to Postgres, so every click is instant.

- layout hosts LiveTournamentProvider (seeded from the single server-side
  getTournament) and is now the only place that fetches; pages became thin
  server shells rendering client views (MatchesView / StandingsView).
- match-card, match-actions, toss-manager, team-score-input-dialog, setup-form,
  sample-results-button and the nav now read/write via the store. All the
  now-pointless useTransition / isSubmitting / isStarting loading state is gone.
- The DB is not written during play yet (Sync/Finish arrive in phase C); the
  per-action server actions in actions.ts are now unused and will be replaced.

Verified: tsc clean, 134/134 tests pass, next build succeeds.
