+++ src/app/tournament/round-robin/[id]/layout.tsx
@@ -5,7 +5,14 @@ import { getTournament } from "@/lib/repositories/tournament-repository";
 import { getTournamentWinner } from "@/contexts/tournament-context/engine";
 import { CompletedBanner } from "@/components/tournaments/completed-banner";
 import { RoundRobinNavigation } from "@/components/tournaments/round-robin-navigation";
+import { LiveTournamentProvider } from "@/contexts/tournament-context/live-provider";
 
+/**
+ * Server shell for a tournament. This is the ONLY place that authenticates and
+ * loads the tournament from the DB; it seeds the live client store, which then
+ * owns all in-progress reads/writes (local-first). The completed banner reflects
+ * the DB (a tournament is only "finished" after Finish & Save).
+ */
 export default async function RoundRobinTournamentLayout({
   children,
   params,
@@ -20,12 +27,16 @@ export default async function RoundRobinTournamentLayout({
 
   return (
     <Box p={{ base: 4, md: 8 }} maxW="600px" mx="auto" w="full">
-      <CompletedBanner
-        completed={record.status === "completed"}
-        winner={getTournamentWinner(record.state)}
-      />
-      <RoundRobinNavigation id={id} isGenerated={record.state.isGenerated} />
-      {children}
+      <LiveTournamentProvider
+        init={{ id: record.id, status: record.status, state: record.state }}
+      >
+        <CompletedBanner
+          completed={record.status === "completed"}
+          winner={getTournamentWinner(record.state)}
+        />
+        <RoundRobinNavigation />
+        {children}
+      </LiveTournamentProvider>
     </Box>
   );
 }
