+++ src/components/tournaments/match-actions.tsx
@@ -9,7 +9,6 @@ import type { Match } from "@/contexts/tournament-context/types";
 interface MatchActionsProps {
   match: Match;
   matchState: string;
-  tournamentId: string;
   onStartMatch: () => void;
   onStartSecondInnings?: () => void;
   onFinishMatch?: () => void;
@@ -18,7 +17,6 @@ interface MatchActionsProps {
 export function MatchActions({
   match,
   matchState,
-  tournamentId,
   onStartMatch,
   onStartSecondInnings,
   onFinishMatch,
@@ -36,7 +34,7 @@ export function MatchActions({
   }
 
   if (matchState === "in-progress-need-toss") {
-    return <TossManager match={match} tournamentId={tournamentId} />;
+    return <TossManager match={match} />;
   }
 
   if (matchState === "first-innings-ready") {
